+++ temba/channels/handlers.py
@@ -25,7 +25,7 @@ from django_redis import get_redis_connection
 from requests import Request
 
 from temba.api.models import WebHookEvent
-from temba.channels.models import Channel, ChannelLog
+from temba.channels.models import Channel, ChannelLog, ChannelEvent
 from temba.contacts.models import Contact, URN
 from temba.flows.models import Flow, FlowRun, FlowStep
 from temba.orgs.models import NEXMO_UUID
@@ -89,6 +89,7 @@ class TwimlAPIHandler(BaseChannelHandler):
 
     def post(self, request, *args, **kwargs):
         from twilio.util import RequestValidator
+        from temba.flows.models import FlowSession
         from temba.msgs.models import Msg
 
         signature = request.META.get('HTTP_X_TWILIO_SIGNATURE', '')
@@ -140,13 +141,14 @@ class TwimlAPIHandler(BaseChannelHandler):
 
                 if flow:
                     call = IVRCall.create_incoming(channel, contact, urn_obj, channel.created_by, call_sid)
+                    session = FlowSession.create(contact, connection=call)
 
                     call.update_status(request.POST.get('CallStatus', None),
                                        request.POST.get('CallDuration', None),
                                        Channel.TYPE_TWILIO)
                     call.save()
 
-                    FlowRun.create(flow, contact.pk, session=call)
+                    FlowRun.create(flow, contact.pk, session=session, connection=call)
                     response = Flow.handle_call(call)
                     return HttpResponse(six.text_type(response))
 
@@ -1072,6 +1074,7 @@ class NexmoCallHandler(BaseChannelHandler):
         return self.get(request, *args, **kwargs)
 
     def get(self, request, *args, **kwargs):
+        from temba.flows.models import FlowSession
         from temba.ivr.models import IVRCall
 
         action = kwargs['action'].lower()
@@ -1119,7 +1122,7 @@ class NexmoCallHandler(BaseChannelHandler):
 
             if call.status == IVRCall.COMPLETED:
                 # if our call is completed, hangup
-                runs = FlowRun.objects.filter(session=call)
+                runs = FlowRun.objects.filter(connection=call)
                 for run in runs:
                     if not run.is_completed():
                         final_step = FlowStep.objects.filter(run=run).order_by('-arrived_on').first()
@@ -1160,8 +1163,9 @@ class NexmoCallHandler(BaseChannelHandler):
 
             if flow:
                 call = IVRCall.create_incoming(channel, contact, urn_obj, channel.created_by, external_id)
+                session = FlowSession.create(contact, connection=call)
 
-                FlowRun.create(flow, contact.pk, session=call)
+                FlowRun.create(flow, contact.pk, session=session, connection=call)
                 response = Flow.handle_call(call)
 
                 event = HttpEvent(request_method, request_path, request_body, 200, six.text_type(response))
@@ -1393,12 +1397,12 @@ class VumiHandler(BaseChannelHandler):
 
                 session_id = str(session_id_part1 + session_id_part2)
 
-                session = USSDSession.handle_incoming(channel=channel, urn=body['from_addr'], content=content,
-                                                      status=status, date=gmt_date, external_id=session_id,
-                                                      message_id=body['message_id'], starcode=body.get('to_addr'))
+                connection = USSDSession.handle_incoming(channel=channel, urn=body['from_addr'], content=content,
+                                                         status=status, date=gmt_date, external_id=session_id,
+                                                         message_id=body['message_id'], starcode=body.get('to_addr'))
 
-                if session:
-                    return HttpResponse("Accepted: %d" % session.id)
+                if connection:
+                    return HttpResponse("Accepted: %d" % connection.id)
                 else:
                     return HttpResponse("Session not handled", status=400)
 
@@ -2105,15 +2109,15 @@ class JunebugHandler(BaseChannelHandler):
                 # Use a session id if provided, otherwise fall back to using the `from` address as the identifier
                 session_id = channel_data.get('session_id') or data['from']
 
-                session = USSDSession.handle_incoming(channel=channel, urn=data['from'], content=data['content'],
-                                                      status=status, date=gmt_date, external_id=session_id,
-                                                      message_id=data['message_id'], starcode=data['to'])
+                connection = USSDSession.handle_incoming(channel=channel, urn=data['from'], content=data['content'],
+                                                         status=status, date=gmt_date, external_id=session_id,
+                                                         message_id=data['message_id'], starcode=data['to'])
 
-                if session:
+                if connection:
                     status = 200
                     response_body = {
                         'status': self.ACK,
-                        'session_id': session.pk,
+                        'session_id': connection.pk,
                     }
                     event = HttpEvent(request_method, request_path, request_body, status, json.dumps(response_body))
                     log_channel(channel, 'Handled USSD message of %s session_event' % (
@@ -2275,7 +2279,9 @@ class JioChatHandler(BaseChannelHandler):
         if msg_type == 'event':
             event = body.get('Event')
             if event == 'subscribe':
-                Trigger.catch_triggers(contact, Trigger.TYPE_FOLLOW, channel)
+                channel_event = ChannelEvent.create(channel, urn, ChannelEvent.TYPE_FOLLOW, timezone.now())
+                channel_event.handle()
+
                 return HttpResponse("New follow handled: %s" % sender_id)
 
         text = ""
@@ -2347,13 +2353,24 @@ class FacebookHandler(BaseChannelHandler):
                 status = []
 
                 for envelope in entry['messaging']:
-                    if 'optin' in envelope:
+                    if 'optin' in envelope or 'referral' in envelope:
                         # check that the recipient is correct for this channel
                         channel_address = str(envelope['recipient']['id'])
                         if channel_address != channel.address:  # pragma: needs cover
                             return HttpResponse("Msg Ignored for recipient id: %s" % channel_address, status=200)
 
-                        referrer_id = envelope['optin'].get('ref')
+                        referrer_id = None
+                        trigger_extra = None
+
+                        if 'optin' in envelope:
+                            referrer_id = envelope['optin'].get('ref')
+                            trigger_extra = envelope['optin']
+                            trigger_extra[ChannelEvent.EXTRA_REFERRER_ID] = referrer_id
+
+                        elif 'referral' in envelope:
+                            referrer_id = envelope['referral'].get('ref')
+                            trigger_extra = envelope['referral']
+                            trigger_extra[ChannelEvent.EXTRA_REFERRER_ID] = referrer_id
 
                         # This is a standard opt in, we know the sender id:
                         #   https://developers.facebook.com/docs/messenger-platform/webhook-reference/optins
@@ -2393,8 +2410,9 @@ class FacebookHandler(BaseChannelHandler):
                             contact = Contact.get_or_create(channel.org, channel.created_by,
                                                             urns=[urn], channel=channel)
 
-                        caught = Trigger.catch_triggers(contact, Trigger.TYPE_REFERRAL, channel,
-                                                        referrer_id=referrer_id, extra=envelope['optin'])
+                        event = ChannelEvent.create(channel, urn, ChannelEvent.TYPE_REFERRAL, timezone.now(),
+                                                    extra=trigger_extra)
+                        caught = event.handle()
 
                         if caught:
                             status.append("Triggered flow for ref: %s" % referrer_id)
@@ -2414,6 +2432,8 @@ class FacebookHandler(BaseChannelHandler):
 
                         content = None
                         postback = None
+                        referrer_id = None
+                        trigger_extra = None
 
                         if 'message' in envelope:
                             if 'text' in envelope['message']:
@@ -2432,6 +2452,10 @@ class FacebookHandler(BaseChannelHandler):
 
                         elif 'postback' in envelope:
                             postback = envelope['postback']['payload']
+                            if 'referral' in envelope['postback']:
+                                trigger_extra = envelope['postback']['referral']
+                                referrer_id = trigger_extra['ref']
+                                trigger_extra[ChannelEvent.EXTRA_REFERRER_ID] = referrer_id
 
                         # if we have some content, load the contact
                         if content or postback:
@@ -2470,9 +2494,18 @@ class FacebookHandler(BaseChannelHandler):
                             Msg.objects.filter(pk=msg.id).update(external_id=envelope['message']['mid'])
                             status.append("Msg %d accepted." % msg.id)
 
+                        # conversation started with a referrer_id that catches a trigger
+                        elif referrer_id:
+                            event = ChannelEvent.create(channel, urn, ChannelEvent.TYPE_REFERRAL, timezone.now(),
+                                                        extra=trigger_extra)
+                            event.handle()
+
+                            status.append("Referral posted with referral id: %s" % referrer_id)
+
                         # a contact pressed "Get Started", trigger any new conversation triggers
                         elif postback == 'get_started':
-                            Trigger.catch_triggers(contact, Trigger.TYPE_NEW_CONVERSATION, channel)
+                            event = ChannelEvent.create(channel, urn, ChannelEvent.TYPE_NEW_CONVERSATION, timezone.now())
+                            event.handle()
                             status.append("Postback handled.")
 
                         else:
@@ -2753,8 +2786,10 @@ class ViberPublicHandler(BaseChannelHandler):
             # }
             viber_id = body['user']['id']
             contact_name = None if channel.org.is_anon else body['user'].get('name')
-            contact = Contact.get_or_create(channel.org, channel.created_by, contact_name, urns=[URN.from_viber(viber_id)])
-            Trigger.catch_triggers(contact, Trigger.TYPE_NEW_CONVERSATION, channel)
+            urn = URN.from_viber(viber_id)
+            contact = Contact.get_or_create(channel.org, channel.created_by, contact_name, urns=[urn])
+            event = ChannelEvent.create(channel, urn, ChannelEvent.TYPE_NEW_CONVERSATION, timezone.now())
+            event.handle()
             return HttpResponse("Subscription for contact: %s handled" % viber_id)
 
         # we currently ignore conversation started events