+++ temba/channels/views.py
@@ -652,7 +652,7 @@ def sync(request, channel_id):
                         if cmd['phone']:
                             urn = URN.from_parts(TEL_SCHEME, cmd['phone'])
                             try:
-                                ChannelEvent.create(channel, urn, cmd['type'], date, duration)
+                                ChannelEvent.create(channel, urn, cmd['type'], date, extra=dict(duration=duration))
                             except ValueError:
                                 # in some cases Android passes us invalid URNs, in those cases just ignore them
                                 pass
@@ -2730,8 +2730,8 @@ class ChannelEventCRUDL(SmartCRUDL):
 
     class Calls(InboxView):
         title = _("Calls")
-        fields = ('contact', 'event_type', 'channel', 'time')
-        default_order = '-time'
+        fields = ('contact', 'event_type', 'channel', 'occurred_on')
+        default_order = '-occurred_on'
         search_fields = ('contact__urns__path__icontains', 'contact__name__icontains')
         system_label = SystemLabel.TYPE_CALLS
         select_related = ('contact', 'channel')
@@ -2759,17 +2759,17 @@ class ChannelLogCRUDL(SmartCRUDL):
             channel = Channel.objects.get(pk=self.request.GET['channel'])
 
             if self.request.GET.get('sessions'):
-                logs = ChannelLog.objects.filter(channel=channel).exclude(session=None).values_list('session_id', flat=True)
+                logs = ChannelLog.objects.filter(channel=channel).exclude(connection=None).values_list('connection_id', flat=True)
                 events = ChannelSession.objects.filter(id__in=logs).order_by('-created_on')
 
                 if self.request.GET.get('errors'):
                     events = events.filter(status=ChannelSession.FAILED)
 
             elif self.request.GET.get('others'):
-                events = ChannelLog.objects.filter(channel=channel, session=None, msg=None).order_by('-created_on')
+                events = ChannelLog.objects.filter(channel=channel, connection=None, msg=None).order_by('-created_on')
 
             else:
-                events = ChannelLog.objects.filter(channel=channel, session=None).exclude(msg=None).order_by('-created_on').select_related('msg__contact', 'msg')
+                events = ChannelLog.objects.filter(channel=channel, connection=None).exclude(msg=None).order_by('-created_on').select_related('msg__contact', 'msg')
                 events.count = lambda: channel.get_non_ivr_log_count()
 
             return events